refactor: extract _printer module from sympy module
